Refactor Post package
