+++ springboot-practice/jefeel/src/main/java/com/practice/jefeel/module/post/controller/dto/PostUpdateRequestDTO.java
@@ -1,6 +1,7 @@
-package com.practice.jefeel.post.controller.dto;
+package com.practice.jefeel.module.post.controller.dto;
 
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,7 +11,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class PostUpdateRequestDTO {
 
+    @NotBlank(message = "제목은 비워둘 수 없습니다")
     private String title;
+    @NotBlank(message = "내용은 비워둘 수 없습니다")
     private String content;
 
     @Builder
