+++ springboot-practice/jefeel/src/main/java/com/practice/jefeel/module/post/domain/repository/PostRepository.java
@@ -1,7 +1,7 @@
-package com.practice.jefeel.post.domain.repository;
+package com.practice.jefeel.module.post.domain.repository;
 
 
-import com.practice.jefeel.post.domain.entity.Post;
+import com.practice.jefeel.module.post.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
